requirements.txt
streamlit
torch
torchvision
pillow
ultralytics
pandas
matplotlib
